Refactor deprecated paper package
