--- src/main/java/codes/wasabi/xclaim/platform/paper/package-info.java
@@ -1,2 +1,0 @@
-@Deprecated
-package codes.wasabi.xclaim.platform.paper;
