+++ src/main/java/codes/wasabi/xclaim/platform/paper_1_17/PaperPlatform.java
@@ -1,4 +1,4 @@
-package codes.wasabi.xclaim.platform.paper;
+package codes.wasabi.xclaim.platform.paper_1_17;
 
 import codes.wasabi.xclaim.XClaim;
 import codes.wasabi.xclaim.platform.Platform;
