+++ src/main/java/codes/wasabi/xclaim/platform/paper_1_17/package-info.java
@@ -1,0 +1,2 @@
+@Deprecated
+package codes.wasabi.xclaim.platform.paper_1_17;
